Features: policy price, next:should review methodology of project managment
There are policy prices linked to the number of night, and their period, and policy prices linked to the number of persons which book
the reservation.
For compute, the booking price, the service will compute with the whole policy prices linked to establishment which is linked to a room.
Actually, i've developped the policy prices that own person as business object.